backend/requirements.txt
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.8.1
attrs==25.3.0
Authlib==1.5.2
bidict==0.23.1
blinker==1.9.0
blis==1.3.3
boto3==1.37.32
botocore==1.37.32
CacheControl==0.14.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.13
dateparser==1.3.0
Django==5.2.1
dnspython==2.7.0
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
eventlet==0.39.1
fastapi==0.123.9
filelock==3.20.0
firebase-admin==6.8.0
Flask==3.1.2
flask-cors==5.0.1
Flask-SocketIO==5.5.1
fpdf==1.7.2
fsspec==2025.10.0
geographiclib==2.0
geopy==2.4.1
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-cloud-texttospeech==2.33.0
google-cloud-translate==3.22.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
googletrans==4.0.0rc1
greenlet==3.2.0
grpc-google-iam-v1==0.14.3
grpcio==1.71.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
libretranslatepy==2.1.1
lxml==6.0.2
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.0
murmurhash==1.0.15
networkx==3.5
numpy==2.3.4
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
preshed==3.0.12
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pymongo==4.12.0
PyMySQL==1.1.1
pyparsing==3.2.3
pypdfium2==5.3.0
PySocks==1.7.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-jose==3.4.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.3
rfc3986==1.5.0
rsa==4.9.1
s3transfer==0.11.4
safetensors==0.6.2
selenium==4.34.0
sentencepiece==0.2.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
smart_open==7.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.2
starlette==0.50.0
sympy==1.14.0
thinc==8.3.10
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
translate==3.8.0
trio==0.30.0
trio-websocket==0.12.2
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.38.0
wasabi==1.1.3
weasel==0.4.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==2.1.1
wsproto==1.2.0
